+++ lib/checkin_scheduler.py
@@ -6,7 +6,7 @@ from typing import TYPE_CHECKING, Any, Dict, List
 from .checkin_handler import CheckInHandler
 from .flight import Flight
 from .log import get_logger
-from .utils import RequestError, make_request
+from .utils import RequestError, get_current_time, make_request
 from .webdriver import WebDriver
 
 if TYPE_CHECKING:
@@ -60,11 +60,12 @@ class CheckInScheduler:
         reservation_info = self._get_reservation_info(confirmation_number)
         logger.debug("%d flights found under current reservation", len(reservation_info))
 
+        current_utc_time = get_current_time()
         flights = []
         # If multiple flights are under the same confirmation number, it will schedule all checkins
         for flight_info in reservation_info:
-            if flight_info["departureStatus"] != "DEPARTED":
-                flight = Flight(flight_info, confirmation_number)
+            flight = Flight(flight_info, confirmation_number)
+            if flight.departure_time > current_utc_time:
                 self._set_same_day_flight(flight, flights)
                 flights.append(flight)
 
@@ -133,13 +134,16 @@ class CheckInScheduler:
         """Remove all scheduled flights that are not in the current flight list"""
         logger.debug("%d flights are currently scheduled. Removing old flights", len(self.flights))
 
+        twenty_four_hr_time = self.reservation_monitor.config.notification_24_hour_time
+
         # Copy the list because it can potentially change inside the loop
         for flight in self.flights[:]:
             if flight not in flights:
                 flight_idx = self.flights.index(flight)
+                flight_time = flight.get_display_time(twenty_four_hr_time)
                 print(
-                    f"Flight from {flight.departure_airport} to {flight.destination_airport} at "
-                    f"{flight.departure_time} UTC is no longer scheduled. Stopping its check-in\n"
+                    f"Flight from {flight.departure_airport} to {flight.destination_airport} on "
+                    f"{flight_time} is no longer scheduled. Stopping its check-in\n"
                 )  # Don't log as it has sensitive information
 
                 self.checkin_handlers[flight_idx].stop_check_in()